refactor: 외부 Client 패키지 분리
